+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 
 ## [Unreleased]
 ### Added
+- Added support for the [Argoverse 2 Motion Forecasting Dataset](https://www.argoverse.org/av2.html#forecasting-link) (see `scenarios/argoverse`)
 ### Changed
 ### Deprecated
 ### Fixed
@@ -19,7 +20,6 @@ Copy and pasting the git commit messages is __NOT__ enough.
 ## [1.0.8] # 2023-03-10
 ### Added
 - Agent manager now has `add_and_emit_social_agent` to generate a new social agent that is immediately in control of a vehicle.
-- Added support for the [Argoverse 2 Motion Forecasting Dataset](https://www.argoverse.org/av2.html#forecasting-link) (see `scenarios/argoverse`)
 ### Changed
 - Changed the minimum supported Python version from 3.7 to 3.8
 ### Deprecated