+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Copy and pasting the git commit messages is __NOT__ enough.
 - Added support for the [Argoverse 2 Motion Forecasting Dataset](https://www.argoverse.org/av2.html#forecasting-link) (see `scenarios/argoverse`)
 - Added `waymo_open_dataset` as a module at the SMARTS repo level, to be able to load waymo scenarios without any external packages
 ### Changed
+- Changed the `lanepoint_spacing` setting in `MapSpec` to be non-optional. Lanepoints are now generated lazily when waypoints are used.
 ### Deprecated
 ### Fixed
 ### Removed